src/backend/teaser/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
bleach==6.0.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
contextlib2==21.6.0
cryptography==39.0.2
distlib==0.3.6
Django==4.1.6
django-admin-honeypot-updated-2021==1.2.0
django-environ==0.9.0
django-extensions==3.2.1
django-ipware==4.0.2
django-ninja==0.21.0
django-ninja-extra==0.18.6
django-prometheus==2.3.1
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
filelock==3.9.0
Flask==2.0.2
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
identify==2.5.18
idna==3.3
injector==0.20.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==6.0.4
mutagen==1.46.0
ninja-schema==0.13.0
nodeenv==1.7.0
numpy==1.21.3
openai==0.27.2
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
platformdirs==3.0.0
praw==7.7.0
prawcore==2.3.0
pre-commit==3.0.4
prometheus-client==0.17.0
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
pyOpenSSL==23.0.0
pyotp==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.5.5
requests==2.26.0
sentry-sdk==1.21.0
six==1.16.0
sqlparse==0.4.3
tinydb==4.7.1
toml==0.10.2
tqdm==4.62.3
tuspy==1.0.0
typing_extensions==4.5.0
tzdata==2022.7
update-checker==0.18.0
urllib3==1.26.15
virtualenv==20.19.0
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.0.2
yarl==1.8.2
yt-dlp==2023.3.4
